Gracefully handle return characters in api token
It's important that the collector doesn't break customer builds. If the
buildkite api token includes a return character (i.e., \r or \n) the
Python libraries throw an InvalidHeader exception, which would cause the
process to error out. This patch gracefully handles the error so the
customer's test suite doesn't fail.