Support failure_expanded
Currently, the entire failure from PyTest (PyTest calls it longreprtext) is
set as failure_reason.

However, failure_reason is intended for a short one-line summary, and
gets truncated to ~1KIB, which isn't great for longreprtext.

Buildkite also supports failure_expanded, to included detailed
information and backtraces on errors.

So, derive that from the various shapes that longrepr can take, and
include them in the uploaded JSON payload.